[wikipedia] cleanup: rename ingest_next to ingest and remove old ingest, also remove old wikipedia/main.ts
