+++ lib/js/wikipedia/ingest.test.mjs
@@ -6,7 +6,7 @@ import {
     semanticMarksFromSchema,
     atomRulesFromSchema,
     markRulesFromSchema,
-} from "./ingest_next";
+} from "./ingest";
 import {
     fromMetaModelJSON,
     ProseMirrorSchemaModel,
@@ -52,7 +52,7 @@ const GENERIC_MARK_RULES = [
     { selector: "a", rule: { kind: "generic", styleName: "link" } },
 ];
 
-// Hard-coded engine test rule tables (do NOT import from ingest_next —
+// Hard-coded engine test rule tables (do NOT import from ingest —
 // these are a moving target; the tests exercise the engine, not Wikipedia
 // parsing policy).
