+++ lib/js/wikipedia/main.mjs
@@ -43,7 +43,7 @@ import { createAndGetDefaultZones } from "../zones.typeroof.jsx";
 // Phase 3: convertDocument kept for operator reference, no longer called.
 // import { convertDocument } from "./ingest";
 
-import { ingestWikipediaDocument } from "./ingest_next";
+import { ingestWikipediaDocument } from "./ingest";
 
 import {
     Controller as TypeStageController,
